refactor:에러처리 훅으로 리팩토링
